convert /api/room/generate endpoint to typescript
